Split SigUtil, Unit etc. for MOBILEAPP too
