+++ wsd/COOLWSD.cpp
@@ -38,10 +38,10 @@
 
 #include <unistd.h>
 #include <sysexits.h>
+#include <sys/resource.h>
+#include <sys/wait.h>
 
 #include <sys/types.h>
-#include <sys/wait.h>
-#include <sys/resource.h>
 
 #include <cassert>
 #include <clocale>
@@ -1282,10 +1282,12 @@ void COOLWSD::setupChildRoot(const bool UseMountNamespaces)
 
 void COOLWSD::innerInitialize(Poco::Util::Application& self)
 {
-    if (!Util::isMobileApp() && geteuid() == 0 && CheckCoolUser)
+#if !MOBILEAPP
+    if (geteuid() == 0 && CheckCoolUser)
     {
         throw std::runtime_error("Do not run as root. Please run as cool user.");
     }
+#endif
 
     Util::setApplicationPath(
         Poco::Path(Poco::Util::Application::instance().commandPath()).parent().toString());
@@ -3507,6 +3509,7 @@ private:
                                           now, *WebServerPoll, factory);
         }
 
+#if !MOBILEAPP
         if (!socket)
         {
             LOG_FTL("Failed to listen on Server port(s) (" << firstPortNumber << '-'
@@ -3514,7 +3517,6 @@ private:
             Util::forcedExit(EX_SOFTWARE);
         }
 
-#if !MOBILEAPP
         LOG_INF('#' << socket->getFD() << " Listening to client connections on port "
                     << ClientPortNumber);
 #else
@@ -4265,11 +4267,11 @@ void dump_state()
 
 #if !MOBILEAPP
     Admin::dumpMetrics();
-#endif
 
     std::lock_guard<std::mutex> docBrokerLock(DocBrokersMutex);
     std::lock_guard<std::mutex> newChildLock(NewChildrenMutex);
     forwardSignal(SIGUSR1);
+#endif
 }
 
 void lslr_childroot()
@@ -4281,6 +4283,7 @@ void lslr_childroot()
 
 void forwardSigUsr2()
 {
+#if !MOBILEAPP
     LOG_TRC("forwardSigUsr2");
 
     if (Util::isKitInProcess())
@@ -4290,6 +4293,7 @@ void forwardSigUsr2()
     std::lock_guard<std::mutex> newChildLock(NewChildrenMutex);
 
     forwardSignal(SIGUSR2);
+#endif
 }
 
 void forwardSignal(const int signum)
@@ -4305,7 +4309,6 @@ void forwardSignal(const int signum)
         LOG_INF("Sending " << name << " to forkit " << COOLWSD::ForKitProcId);
         ::kill(COOLWSD::ForKitProcId, signum);
     }
-#endif
 
     for (const auto& child : NewChildren)
     {
@@ -4325,6 +4328,7 @@ void forwardSignal(const int signum)
             ::kill(docBroker->getPid(), signum);
         }
     }
+#endif
 }
 
 // Avoid this in the Util::isFuzzing() case because libfuzzer defines its own main().
